fix member adapter accessor test.
The test uses the firstname as test case, but the firstname is not part
of the address and cannot be changed in the member block and therefore
is not the right attribute.
It is changed to the `address` attribute.

This fixes test failures which are happing since the builder no longer
blindly sets attributes:

https://github.com/4teamwork/ftw.builder/pull/45