Fix broken contentpage with memberblock
We disabled dizo for memberblocks so that it does not throw a ecoding error in tests. But it had the side effect to break the whole content page it was in. To fix this we just rewrite the tests to not get in the situation where we get the encoding error and do not disable diazo.